annoying
the time limit error was annoying because i was doing it with time complexity of O(N log n) which was most fitted for the problem.  I was trying to use my own lessNum  function which was equivalent to min() function. Using my function own function was the cause for time limit exceed error.